refactor: extract session fork logic
