+++ src/acp-agent.ts
@@ -61,7 +61,6 @@ import {
   EffortLevel,
   FastModeDisabledReason,
   FastModeState,
-  forkSession as forkClaudeSession,
   getSessionMessages,
   listSessions,
   McpServerConfig,
@@ -119,6 +118,7 @@ import {
   refusalFallbackResultFromResponse,
   refusalFallbackToCreateRequest,
 } from "./elicitation.js";
+import { forkSession } from "./fork-session.js";
 import { ALLOW_BYPASS, resolvePermissionMode } from "./permissions/modes.js";
 import { normalizeDurablePermissionChangeSet } from "./permissions/normalization.js";
 import { buildClaudePermissionOptions } from "./permissions/options.js";
@@ -270,31 +270,6 @@ type SteerMeta = {
     idleBehavior?: "promptRequired";
   };
 };
-
-type ForkSessionMeta = {
-  [key: string]: unknown;
-  jetbrains?: {
-    air?: {
-      fork?: {
-        version?: number;
-        messageId?: string;
-      };
-    };
-  };
-};
-
-function forkPointMessageId(meta: unknown): string | undefined {
-  const fork = (meta as ForkSessionMeta | null | undefined)?.jetbrains?.air?.fork;
-  if (fork?.version !== 1) return undefined;
-  const messageId = fork.messageId?.trim();
-  return messageId || undefined;
-}
-
-function forkPointMessageIdCandidates(messageId: string): string[] {
-  // Older AIR builds sent their visible segment id. Prefer the exact id before its ACP source id.
-  const protocolMessageId = messageId.replace(/:segment:\d+$/, "");
-  return protocolMessageId === messageId ? [messageId] : [messageId, protocolMessageId];
-}
 
 /** Params of a {@link STEER_METHOD} request. Shaped like the relevant subset of
  *  a `PromptRequest` so the same `promptToClaude` conversion applies. Delivery
@@ -1665,38 +1640,10 @@ export class ClaudeAcpAgent {
 
   async unstable_forkSession(params: ForkSessionRequest): Promise<ForkSessionResponse> {
     if (this.providerUpdate) await this.providerUpdate;
-    const messageId = forkPointMessageId(params._meta);
-    if (messageId) {
-      const candidateIds = forkPointMessageIdCandidates(messageId);
-      const liveMappings = this.sessions[params.sessionId]?.messageIdToUuid;
-      const liveUuid = candidateIds
-        .map((candidateId) => liveMappings?.get(candidateId))
-        .find(Boolean);
-      const history = liveUuid
-        ? undefined
-        : await getSessionMessages(params.sessionId, { dir: params.cwd });
-      const messageUuid =
-        liveUuid ??
-        candidateIds
-          .map(
-            (candidateId) =>
-              history?.find((message) => messageIdForGrouping(message) === candidateId)?.uuid,
-          )
-          .find(Boolean);
-      if (!messageUuid) {
-        throw RequestError.invalidParams(
-          { messageId },
-          `Fork point message ${messageId} was not found in session ${params.sessionId}`,
-        );
-      }
-      const forked = await forkClaudeSession(params.sessionId, {
-        dir: params.cwd,
-        upToMessageId: messageUuid,
-      });
-      return { sessionId: forked.sessionId };
-    }
-    const forked = await forkClaudeSession(params.sessionId, { dir: params.cwd });
-    return { sessionId: forked.sessionId };
+    return forkSession(params, {
+      liveMessageIdToUuid: this.sessions[params.sessionId]?.messageIdToUuid,
+      messageIdForGrouping,
+    });
   }
 
   async resumeSession(params: ResumeSessionRequest): Promise<ResumeSessionResponse> {
